# Build a local memSeries cache while replaying WLs