This is an attempt to access a smaller data structure without locking, but also without refactoring a lot of code.