@colega 2024-08-29
`BenchmarkLoadWLs` results:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
                                                                                                                                                             │     main     │                cache                │
                                                                                                                                                             │    sec/op    │   sec/op     vs base                │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         164.6m ±  1%   165.8m ± 5%        ~ (p=0.123 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        167.1m ±  2%   167.6m ± 1%        ~ (p=0.529 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        169.9m ±  2%   170.6m ± 3%        ~ (p=0.165 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12       196.3m ±  1%   197.2m ± 0%        ~ (p=0.165 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         223.8m ±  3%   188.7m ± 2%  -15.69% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         246.1m ±  6%   209.1m ± 2%  -15.02% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         78.98m ±  1%   77.82m ± 1%   -1.47% (p=0.002 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         81.40m ±  5%   80.19m ± 1%   -1.48% (p=0.001 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         83.38m ±  1%   83.09m ± 1%        ~ (p=0.218 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        98.85m ±  0%   98.91m ± 4%        ~ (p=0.796 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     826.0m ±  1%   799.5m ± 2%   -3.21% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     849.5m ±  1%   820.9m ± 2%   -3.36% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     874.5m ± 19%   850.4m ± 9%   -2.76% (p=0.009 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     1.079 ±  1%    1.057 ± 8%   -2.04% (p=0.023 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        279.8m ±  1%   278.3m ± 0%   -0.55% (p=0.023 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        282.6m ±  1%   280.5m ± 1%   -0.76% (p=0.043 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        286.0m ±  1%   283.2m ± 1%   -1.00% (p=0.003 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12       301.1m ±  1%   299.8m ± 1%        ~ (p=0.105 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         251.1m ±  1%   249.2m ± 1%   -0.77% (p=0.023 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         253.9m ±  1%   251.2m ± 2%   -1.07% (p=0.011 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         254.9m ±  7%   266.2m ± 7%        ~ (p=0.190 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12        268.7m ±  5%   265.7m ± 2%        ~ (p=0.165 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12     3.289 ±  1%    3.244 ± 1%   -1.36% (p=0.004 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12     3.311 ±  1%    3.292 ± 1%   -0.57% (p=0.029 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12     3.371 ±  2%    3.282 ± 1%   -2.62% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    3.516 ±  1%    3.471 ± 1%   -1.27% (p=0.000 n=10)
geomean                                                                                                                                                        372.0m         364.4m        -2.04%

                                                                                                                                                             │     main     │                cache                 │
                                                                                                                                                             │     B/op     │     B/op      vs base                │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         36.58Mi ± 0%   36.64Mi ± 3%   +0.17% (p=0.001 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        38.89Mi ± 1%   39.44Mi ± 2%   +1.42% (p=0.002 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        41.04Mi ± 0%   41.12Mi ± 0%   +0.19% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12       56.33Mi ± 0%   56.38Mi ± 0%   +0.09% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         195.6Mi ± 2%   195.2Mi ± 1%        ~ (p=0.796 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         231.6Mi ± 2%   225.7Mi ± 3%   -2.56% (p=0.011 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         53.91Mi ± 1%   54.55Mi ± 2%   +1.18% (p=0.023 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         55.99Mi ± 1%   57.17Mi ± 1%   +2.10% (p=0.001 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         57.86Mi ± 1%   59.04Mi ± 1%   +2.04% (p=0.001 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        67.20Mi ± 1%   68.30Mi ± 2%   +1.64% (p=0.043 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     260.3Mi ± 1%   302.2Mi ± 1%  +16.08% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     270.6Mi ± 1%   312.4Mi ± 1%  +15.45% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     285.6Mi ± 7%   327.5Mi ± 1%  +14.68% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12    387.7Mi ± 1%   429.7Mi ± 1%  +10.84% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        541.8Mi ± 0%   543.4Mi ± 0%        ~ (p=0.143 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        543.9Mi ± 0%   545.4Mi ± 0%   +0.27% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        545.6Mi ± 0%   546.9Mi ± 0%   +0.24% (p=0.019 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12       554.5Mi ± 0%   556.1Mi ± 0%   +0.30% (p=0.015 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         462.2Mi ± 0%   462.6Mi ± 0%        ~ (p=0.353 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         462.9Mi ± 0%   464.8Mi ± 0%   +0.40% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         463.9Mi ± 0%   466.0Mi ± 0%   +0.46% (p=0.002 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12        472.3Mi ± 0%   472.6Mi ± 0%        ~ (p=0.247 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    4.197Gi ± 0%   4.237Gi ± 0%   +0.97% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    4.207Gi ± 0%   4.244Gi ± 0%   +0.88% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    4.220Gi ± 0%   4.257Gi ± 0%   +0.88% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12   4.305Gi ± 0%   4.344Gi ± 0%   +0.90% (p=0.000 n=10)
geomean                                                                                                                                                        297.3Mi        304.8Mi        +2.53%

                                                                                                                                                             │    main     │               cache                │
                                                                                                                                                             │  allocs/op  │  allocs/op   vs base               │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         392.8k ± 0%   392.9k ± 2%  +0.03% (p=0.001 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        505.2k ± 1%   509.2k ± 2%  +0.81% (p=0.011 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        617.5k ± 0%   617.6k ± 0%  +0.01% (p=0.004 n=10)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12       1.484M ± 0%   1.485M ± 0%  +0.01% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         2.009M ± 0%   2.008M ± 0%       ~ (p=0.579 n=10)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         2.609M ± 0%   2.608M ± 0%       ~ (p=0.052 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         401.5k ± 0%   402.5k ± 0%  +0.24% (p=0.019 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         462.3k ± 0%   462.2k ± 0%       ~ (p=0.684 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12         553.6k ± 0%   553.0k ± 0%       ~ (p=0.436 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12        1.125M ± 0%   1.125M ± 0%       ~ (p=0.315 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     2.964M ± 0%   2.981M ± 0%  +0.59% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     3.563M ± 0%   3.580M ± 0%  +0.48% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12     4.465M ± 3%   4.482M ± 0%  +0.39% (p=0.019 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-12    10.31M ± 0%   10.33M ± 0%  +0.18% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        2.862M ± 0%   2.863M ± 0%       ~ (p=0.684 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        2.922M ± 0%   2.923M ± 0%       ~ (p=0.190 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12        3.012M ± 0%   3.012M ± 0%       ~ (p=0.912 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-12       3.582M ± 0%   3.583M ± 0%       ~ (p=0.280 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         2.693M ± 0%   2.692M ± 0%       ~ (p=0.684 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         2.727M ± 1%   2.753M ± 0%  +0.95% (p=0.000 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12         2.790M ± 2%   2.842M ± 2%       ~ (p=0.123 n=10)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-12        3.359M ± 0%   3.359M ± 0%       ~ (p=0.739 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    23.74M ± 0%   23.76M ± 0%  +0.06% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    24.34M ± 0%   24.36M ± 0%  +0.06% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12    25.24M ± 0%   25.25M ± 0%  +0.04% (p=0.000 n=10)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-12   30.94M ± 0%   30.96M ± 0%  +0.05% (p=0.000 n=10)
geomean                                                                                                                                                        2.691M        2.697M       +0.22%
```

Pending to benchmark on a real WAL